DTD-4488 – Cucumber to ScalaTest
Corrected spacing throughout.
Moved response data definitions out of the methods in all scenarios and passed the val name into the method for consistency across the repository.
Removed unnecessary helpers.
Used Ctrl + Option + O to optimise and rearrange imports.
Removed outcome where it was inferr
